feat: adds state change cause to channel
This adds a `state_change` 'cause' to a channel.
A 'cause' is some initial 'reason' a channel was created or closed by:

 - REASON_UNKOWN  (no reason given, use prior cause if known)
 - REASON_LOCAL   (bad HTLC, broken crypto, peer does fishy things ...)
 - REASON_USER    (by 'hand' or plugin)
 - REASON_REMOTE  (bilateral close initiated by remote peer)
 - REASON_ONCHAIN (e.g. channel got settled while we were offline)

If a 'cause' is known and a subsequent state change is made with
`REASON_UNKNOWN` the preceding cause will be used as reason, since a lot
(all `REASON_UNKNOWN`) state changes are a subsequent consequences of a prior
cause (LOCAL, USER, REMOTE, ONCHAIN).

Changelog-Added: Plugins: Channel closure resaon/cause to channel_state_changed notification